+++ processor/src/main/java/org/realityforge/arez/processor/AutorunDescriptor.java
@@ -8,6 +8,7 @@ import com.squareup.javapoet.TypeSpec;
 import java.util.ArrayList;
 import java.util.Objects;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.type.ExecutableType;
@@ -16,6 +17,7 @@ import javax.lang.model.type.TypeMirror;
 /**
  * The class that represents the parsed state of @Autorun methods on a @ArezComponent annotated class.
  */
+@SuppressWarnings( "Duplicates" )
 final class AutorunDescriptor
 {
   @Nonnull
@@ -27,6 +29,10 @@ final class AutorunDescriptor
   private final ExecutableElement _autorun;
   @Nonnull
   private final ExecutableType _autorunType;
+  @Nullable
+  private ExecutableElement _refMethod;
+  @Nullable
+  private ExecutableType _refMethodType;
 
   AutorunDescriptor( @Nonnull final ComponentDescriptor componentDescriptor,
                      @Nonnull final String name,
@@ -45,6 +51,25 @@ final class AutorunDescriptor
   String getName()
   {
     return _name;
+  }
+
+  boolean hasRefMethod()
+  {
+    return null != _refMethod;
+  }
+
+  @Nonnull
+  ExecutableElement getRefMethod()
+    throws ArezProcessorException
+  {
+    assert null != _refMethod;
+    return _refMethod;
+  }
+
+  void setRefMethod( @Nonnull final ExecutableElement method, @Nonnull final ExecutableType methodType )
+  {
+    _refMethod = Objects.requireNonNull( method );
+    _refMethodType = Objects.requireNonNull( methodType );
   }
 
   @Nonnull
@@ -105,6 +130,34 @@ final class AutorunDescriptor
     throws ArezProcessorException
   {
     builder.addMethod( buildAutorun() );
+    if ( null != _refMethod )
+    {
+      builder.addMethod( buildObserverRefMethod() );
+    }
+  }
+
+  /**
+   * Generate the accessor for ref method.
+   */
+  @Nonnull
+  private MethodSpec buildObserverRefMethod()
+    throws ArezProcessorException
+  {
+    assert null != _refMethod;
+    assert null != _refMethodType;
+    final MethodSpec.Builder builder = MethodSpec.methodBuilder( _refMethod.getSimpleName().toString() );
+    ProcessorUtil.copyAccessModifiers( _refMethod, builder );
+    ProcessorUtil.copyTypeParameters( _refMethodType, builder );
+    ProcessorUtil.copyDocumentedAnnotations( _refMethod, builder );
+
+    builder.addAnnotation( Override.class );
+    builder.returns( TypeName.get( _refMethodType.getReturnType() ) );
+
+    GeneratorUtil.generateNotDisposedInvariant( _componentDescriptor, builder );
+
+    builder.addStatement( "return $N", GeneratorUtil.FIELD_PREFIX + getName() );
+
+    return builder.build();
   }
 
   /**