+++ processor/src/main/java/org/realityforge/arez/processor/TrackedDescriptor.java
@@ -41,6 +41,10 @@ final class TrackedDescriptor
   private ExecutableType _trackedMethodType;
   @Nullable
   private ExecutableElement _onDepsChangedMethod;
+  @Nullable
+  private ExecutableElement _refMethod;
+  @Nullable
+  private ExecutableType _refMethodType;
 
   TrackedDescriptor( @Nonnull final ComponentDescriptor componentDescriptor, @Nonnull final String name )
   {
@@ -52,6 +56,25 @@ final class TrackedDescriptor
   String getName()
   {
     return _name;
+  }
+
+  boolean hasRefMethod()
+  {
+    return null != _refMethod;
+  }
+
+  @Nonnull
+  ExecutableElement getRefMethod()
+    throws ArezProcessorException
+  {
+    assert null != _refMethod;
+    return _refMethod;
+  }
+
+  void setRefMethod( @Nonnull final ExecutableElement method, @Nonnull final ExecutableType methodType )
+  {
+    _refMethod = Objects.requireNonNull( method );
+    _refMethodType = Objects.requireNonNull( methodType );
   }
 
   @Nonnull
@@ -168,6 +191,34 @@ final class TrackedDescriptor
     throws ArezProcessorException
   {
     builder.addMethod( buildTracked() );
+    if ( null != _refMethod )
+    {
+      builder.addMethod( buildObserverRefMethod() );
+    }
+  }
+
+  /**
+   * Generate the accessor for ref method.
+   */
+  @Nonnull
+  private MethodSpec buildObserverRefMethod()
+    throws ArezProcessorException
+  {
+    assert null != _refMethod;
+    assert null != _refMethodType;
+    final MethodSpec.Builder builder = MethodSpec.methodBuilder( _refMethod.getSimpleName().toString() );
+    ProcessorUtil.copyAccessModifiers( _refMethod, builder );
+    ProcessorUtil.copyTypeParameters( _refMethodType, builder );
+    ProcessorUtil.copyDocumentedAnnotations( _refMethod, builder );
+
+    builder.addAnnotation( Override.class );
+    builder.returns( TypeName.get( _refMethodType.getReturnType() ) );
+
+    GeneratorUtil.generateNotDisposedInvariant( _componentDescriptor, builder );
+
+    builder.addStatement( "return $N", GeneratorUtil.FIELD_PREFIX + getName() );
+
+    return builder.build();
   }
 
   /**