+++ processor/src/test/resources/input/com/example/observer_ref/RefOnBothModel.java
@@ -1,0 +1,39 @@
+package com.example.observer_ref;
+
+import org.realityforge.arez.Observer;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Autorun;
+import org.realityforge.arez.annotations.ObserverRef;
+import org.realityforge.arez.annotations.OnDepsChanged;
+import org.realityforge.arez.annotations.Track;
+
+@ArezComponent
+public class RefOnBothModel
+{
+  @Track
+  public void render( final long time, float someOtherParameter )
+  {
+  }
+
+  @OnDepsChanged
+  public void onRenderDepsChanged()
+  {
+  }
+
+  @ObserverRef
+  Observer getRenderObserver()
+  {
+    throw new IllegalStateException();
+  }
+
+  @Autorun
+  public void doStuff()
+  {
+  }
+
+  @ObserverRef
+  Observer getDoStuffObserver()
+  {
+    throw new IllegalStateException();
+  }
+}