+++ processor/src/test/resources/input/com/example/observer_ref/RefOnTrackedModel.java
@@ -1,0 +1,27 @@
+package com.example.observer_ref;
+
+import org.realityforge.arez.Observer;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.ObserverRef;
+import org.realityforge.arez.annotations.OnDepsChanged;
+import org.realityforge.arez.annotations.Track;
+
+@ArezComponent
+public class RefOnTrackedModel
+{
+  @Track
+  public void render( final long time, float someOtherParameter )
+  {
+  }
+
+  @OnDepsChanged
+  public void onRenderDepsChanged()
+  {
+  }
+
+  @ObserverRef
+  Observer getRenderObserver()
+  {
+    throw new IllegalStateException();
+  }
+}